pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelforge-lite"
version = "0.1.0"
description = "Pixel-art-only image-to-pixel-art conversion SDK (lite variant)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "arki" }]
keywords = ["pixel art", "image processing", "quantization", "retro"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.23",
    "Pillow>=9.0",
    "opencv-python>=4.5",
]

[project.optional-dependencies]
# Development dependencies are kept for maintainer convenience; the test and
# benchmark suites themselves live in a separate private workspace.
test = ["pytest>=7.0", "pytest-cov>=4.0"]
bench = ["pytest-benchmark>=4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pixelforge_litemodel*"]

[tool.setuptools.package-data]
pixelforge_litemodel = ["palettes/data/*.json"]
